Move pkg directory.
