+++ empire/server/heroku/authentication_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/remind101/empire/empire"
-	"github.com/remind101/empire/pkg/httpx"
+	"github.com/remind101/empire/empire/pkg/httpx"
 	"golang.org/x/net/context"
 )
